+++ tests/Unit/ThemePolicyTest.php
@@ -10,7 +10,7 @@ use Illuminate\Support\Facades\Config;
 test('it allows viewAny by default if authorization is disabled', function () {
     Config::set('themer-luncher.authorization.enabled', false);
 
-    $policy = new ThemePolicy;
+    $policy = new ThemePolicy();
     expect($policy->viewAny(null))->toBeTrue();
 });
 
@@ -18,7 +18,7 @@ test('it respects config permissions', function () {
     Config::set('themer-luncher.authorization.enabled', true);
     Config::set('themer-luncher.authorization.permissions.viewAny', true);
 
-    $policy = new ThemePolicy;
+    $policy = new ThemePolicy();
     expect($policy->viewAny(null))->toBeTrue();
 
     Config::set('themer-luncher.authorization.permissions.viewAny', false);